Added more possible commands and changed the data type to an int

+++ MonoGame/Networking/Client.cs
@@ -84,7 +84,7 @@ public class Client : IDisposable
         
         AddHeaders(ControlDataType, writer);
         
-        writer.Write((byte)data);
+        writer.Write((int)data);
 
         _udpClient.Send(ms.GetBuffer(), (int)ms.Position);
     }
